feat: adding cached functions to ssr
