+++ app/community/[communityId]/layout.tsx
@@ -1,13 +1,11 @@
 import { CommunityPageNavigator } from "@/components/Pages/Communities/CommunityPageNavigator";
-import { zeroUID } from "@/utilities/commons";
 import { communityColors } from "@/utilities/communityColors";
 import { envVars } from "@/utilities/enviromentVars";
-import { gapIndexerApi } from "@/utilities/gapIndexerApi";
 import { defaultMetadata } from "@/utilities/meta";
 import { pagesOnRoot } from "@/utilities/pagesOnRoot";
+import { getCommunityData } from "@/utilities/queries/getCommunityData";
 import { ICommunityResponse } from "@show-karma/karma-gap-sdk/core/class/karma-indexer/api/types";
 import { Metadata } from "next";
-import { notFound } from "next/navigation";
 import { CommunityImpactStatCards } from "@/components/Pages/Communities/Impact/StatCards";
 
 type Props = {
@@ -18,17 +16,9 @@ type Props = {
 export async function generateMetadata(props: Props): Promise<Metadata> {
   const params = await props.params;
   const communityId = params.communityId;
-  let communityName = communityId;
-
-  try {
-    const { data } = await gapIndexerApi.communityBySlug(communityId);
-    communityName = data?.details?.data?.name || communityId;
-    if (!data || data?.uid === zeroUID || !data?.details?.data?.name) {
-      notFound();
-    }
-  } catch {
-    notFound();
-  }
+  
+  const community = await getCommunityData(communityId);
+  const communityName = community?.details?.data?.name || communityId;
 
   const dynamicMetadata = {
     title: `Karma GAP - ${communityName} community grants`,
@@ -81,29 +71,12 @@ export default async function Layout(
   } = props;
 
   const { communityId } = params;
-  let community: ICommunityResponse | null = null;
 
   if (pagesOnRoot.includes(communityId)) {
     return undefined;
   }
 
-  await Promise.all(
-    [
-      async () => {
-        try {
-          const { data } = await gapIndexerApi.communityBySlug(communityId);
-          community = data as ICommunityResponse;
-        } catch {
-          console.log("Not found community", communityId);
-          community = null;
-        }
-      },
-    ].map((func) => func())
-  );
-
-  if (!community) {
-    notFound();
-  }
+  const community = await getCommunityData(communityId);
 
   return (
     <div className="flex w-full h-full max-w-full flex-col justify-start max-lg:flex-col">
